fix(sql): termo não-recursivo do fold precisa ser bigint
row_number() devolve bigint; `0 AS ord` entrava como integer e o
Postgres recusava a CTE recursiva (42804).

+++ supabase/migrations/APPLY_MANUALLY_retag_variantes_schema.sql
@@ -180,7 +180,9 @@ base AS (
 
 -- Aplica os renames em cascata, um por passo.
 fold AS (
-  SELECT id, 0 AS ord, html, tagged FROM base
+  -- 0::bigint: `row_number()` devolve bigint e o Postgres exige que o
+  -- termo não-recursivo tenha o MESMO tipo do recursivo.
+  SELECT id, 0::bigint AS ord, html, tagged FROM base
   UNION ALL
   SELECT
     f.id,
